+++ app/src/main/java/woowacourse/movie/model/MovieUiModel.kt
@@ -1,11 +1,12 @@
 package woowacourse.movie.model
 
 import androidx.annotation.DrawableRes
+import com.example.domain.model.PlayingTimes
 
 data class MovieUiModel(
     @DrawableRes val image: Int,
     val title: String,
-    val playingTimes: com.example.domain.model.PlayingTimes,
+    val playingTimes: PlayingTimes,
     val runningTime: Int,
     val description: String
-) : java.io.Serializable
+)
