+++ app/src/test/java/woowacourse/movie/MovieUiModelDayPolicyTest.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie
 
-import com.example.domain.model.Price
+import com.example.domain.model.price.Price
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
 import java.time.LocalDate
