+++ app/src/test/java/woowacourse/movie/NightPolicyTest.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie
 
-import com.example.domain.model.Price
 import com.example.domain.model.policy.MorningPolicy
+import com.example.domain.model.price.Price
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
 import java.time.LocalDate
