+++ domain/src/main/java/com/example/domain/model/policy/MovieDayPolicy.kt
@@ -1,11 +1,12 @@
 package com.example.domain.model.policy
 
+import com.example.domain.model.price.Price
 import java.time.LocalDate
 import java.time.LocalTime
 
 class MovieDayPolicy : DiscountPolicy() {
-    override fun getDiscountPrice(price: com.example.domain.model.Price): com.example.domain.model.Price {
-        return com.example.domain.model.Price((price.price * 0.9).toInt())
+    override fun getDiscountPrice(price: Price): Price {
+        return Price((price.price * 0.9).toInt())
     }
 
     override fun isAvailable(date: LocalDate, time: LocalTime): Boolean {
