+++ domain/src/main/java/com/example/domain/model/price/Price.kt
@@ -1,4 +1,4 @@
-package com.example.domain.model
+package com.example.domain.model.price
 
 @JvmInline
 value class Price(val price: Int = DEFAULT) {
